+++ tests/Integration/Fixtures/FeedFixture.php
@@ -12,7 +12,6 @@
 
 namespace OCA\News\Tests\Integration\Fixtures;
 
-
 use OCA\News\Db\Feed;
 
 class FeedFixture extends Feed
@@ -20,31 +19,33 @@ class FeedFixture extends Feed
 
     use Fixture;
 
-    public function __construct(array $defaults=[])
+    public function __construct(array $defaults = [])
     {
+        parent::__construct();
         $defaults = array_merge(
             [
-            'userId' => 'test',
-            'urlHash' => 'urlHash',
-            'url' => 'http://the.url.com',
-            'title' => 'title',
-            'faviconLink' => 'http://feed.com/favicon.ico',
-            'added' => 3000,
-            'folderId' => null,
-            'link' => 'http://feed.com/rss',
-            'preventUpdate' => false,
-            'deletedAt' => 0,
-            'articlesPerUpdate' => 40,
-            'httpLastModified' => 10,
-            'httpEtag' => '',
-            'location' => 'http://feed.com/rss',
-            'ordering' => 0,
-            'fullTextEnabled' => false,
-            'pinned' => false,
-            'updateMode' => 0,
-            'updateErrorCount' => 0,
-            'lastUpdateError' => '',
-            ], $defaults
+                'userId' => 'test',
+                'urlHash' => 'urlHash',
+                'url' => 'http://the.url.com',
+                'title' => 'title',
+                'faviconLink' => 'http://feed.com/favicon.ico',
+                'added' => 3000,
+                'folderId' => null,
+                'link' => 'http://feed.com/rss',
+                'preventUpdate' => false,
+                'deletedAt' => 0,
+                'articlesPerUpdate' => 40,
+                'httpLastModified' => 10,
+                'httpEtag' => '',
+                'location' => 'http://feed.com/rss',
+                'ordering' => 0,
+                'fullTextEnabled' => false,
+                'pinned' => false,
+                'updateMode' => 0,
+                'updateErrorCount' => 0,
+                'lastUpdateError' => '',
+            ],
+            $defaults
         );
         unset($defaults['items']);
         $this->fillDefaults($defaults);